refactor: useProducts, useNotifications 파일 분리
