+++ src/basic/App.tsx
@@ -7,20 +7,25 @@ import Toast from './components/ui/Toast';
 import Header from './components/ui/Header';
 import { initialProducts } from './data/mockProducts';
 import { initialCoupons } from './data/mockCoupons';
+import { useProducts } from './hooks/product/useProducts';
+import { useNotifications } from './hooks/notifications/useNotifications';
 
 
 const App = () => {
-  const [products, setProducts] = useState<ProductWithUI[]>(() => {
-    const saved = localStorage.getItem('products');
-    if (saved) {
-      try {
-        return JSON.parse(saved);
-      } catch {
-        return initialProducts;
-      }
-    }
-    return initialProducts;
-  });
+  // const [products, setProducts] = useState<ProductWithUI[]>(() => {
+  //   const saved = localStorage.getItem('products');
+  //   if (saved) {
+  //     try {
+  //       return JSON.parse(saved);
+  //     } catch {
+  //       return initialProducts;
+  //     }
+  //   }
+  //   return initialProducts;
+  // });
+
+  const {products, setProducts, addProduct, updateProduct, deleteProduct} = useProducts()
+  const {notifications, setNotifications, addNotification} = useNotifications()
 
   const [cart, setCart] = useState<CartItem[]>(() => {
     const saved = localStorage.getItem('cart');
@@ -48,7 +53,7 @@ const App = () => {
 
   const [selectedCoupon, setSelectedCoupon] = useState<Coupon | null>(null);
   const [isAdmin, setIsAdmin] = useState(false);
-  const [notifications, setNotifications] = useState<Notification[]>([]);
+  // const [notifications, setNotifications] = useState<Notification[]>([]);
   const [showCouponForm, setShowCouponForm] = useState(false);
   const [activeTab, setActiveTab] = useState<'products' | 'coupons'>('products');
   const [showProductForm, setShowProductForm] = useState(false);
@@ -149,17 +154,17 @@ const App = () => {
     return remaining;
   };
 
-  const addNotification = useCallback(
-    (message: string, type: 'error' | 'success' | 'warning' = 'success') => {
-      const id = Date.now().toString();
-      setNotifications((prev) => [...prev, { id, message, type }]);
+  // const addNotification = useCallback(
+  //   (message: string, type: 'error' | 'success' | 'warning' = 'success') => {
+  //     const id = Date.now().toString();
+  //     setNotifications((prev) => [...prev, { id, message, type }]);
 
-      setTimeout(() => {
-        setNotifications((prev) => prev.filter((n) => n.id !== id));
-      }, 3000);
-    },
-    [],
-  );
+  //     setTimeout(() => {
+  //       setNotifications((prev) => prev.filter((n) => n.id !== id));
+  //     }, 3000);
+  //   },
+  //   [],
+  // );
 
   const [totalCartItem, setTotalCartItem] = useState(0);
 
@@ -274,35 +279,35 @@ const App = () => {
     setSelectedCoupon(null);
   }, [addNotification]);
 
-  const addProduct = useCallback(
-    (newProduct: Omit<ProductWithUI, 'id'>) => {
-      const product: ProductWithUI = {
-        ...newProduct,
-        id: `p${Date.now()}`,
-      };
-      setProducts((prev) => [...prev, product]);
-      addNotification('상품이 추가되었습니다.', 'success');
-    },
-    [addNotification],
-  );
+  // const addProduct = useCallback(
+  //   (newProduct: Omit<ProductWithUI, 'id'>) => {
+  //     const product: ProductWithUI = {
+  //       ...newProduct,
+  //       id: `p${Date.now()}`,
+  //     };
+  //     setProducts((prev) => [...prev, product]);
+  //     addNotification('상품이 추가되었습니다.', 'success');
+  //   },
+  //   [addNotification],
+  // );
 
-  const updateProduct = useCallback(
-    (productId: string, updates: Partial<ProductWithUI>) => {
-      setProducts((prev) =>
-        prev.map((product) => (product.id === productId ? { ...product, ...updates } : product)),
-      );
-      addNotification('상품이 수정되었습니다.', 'success');
-    },
-    [addNotification],
-  );
+  // const updateProduct = useCallback(
+  //   (productId: string, updates: Partial<ProductWithUI>) => {
+  //     setProducts((prev) =>
+  //       prev.map((product) => (product.id === productId ? { ...product, ...updates } : product)),
+  //     );
+  //     addNotification('상품이 수정되었습니다.', 'success');
+  //   },
+  //   [addNotification],
+  // );
 
-  const deleteProduct = useCallback(
-    (productId: string) => {
-      setProducts((prev) => prev.filter((p) => p.id !== productId));
-      addNotification('상품이 삭제되었습니다.', 'success');
-    },
-    [addNotification],
-  );
+  // const deleteProduct = useCallback(
+  //   (productId: string) => {
+  //     setProducts((prev) => prev.filter((p) => p.id !== productId));
+  //     addNotification('상품이 삭제되었습니다.', 'success');
+  //   },
+  //   [addNotification],
+  // );
 
   const addCoupon = useCallback(
     (newCoupon: Coupon) => {
