+++ src/basic/hooks/notifications/useNotifications.ts
@@ -1,0 +1,25 @@
+import { useCallback, useState } from 'react';
+import { Notification } from '../../../types';
+
+export const useNotifications = () => {
+
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: 'error' | 'success' | 'warning' = 'success') => {
+      const id = Date.now().toString();
+      setNotifications((prev) => [...prev, { id, message, type }]);
+
+      setTimeout(() => {
+        setNotifications((prev) => prev.filter((n) => n.id !== id));
+      }, 3000);
+    },
+    [],
+  );
+
+  return {
+    notifications,
+    setNotifications,
+    addNotification
+  }
+}
